Validate Authentication while webhook creation and deletion

+++ app/Http/Livewire/User/Settings/Integrations.php
@@ -33,26 +33,30 @@ class Integrations extends Component
         }
 
         if (Auth::check()) {
-            $this->validate([
-                'name' => 'required|min:2|max:20',
-                'product' => 'nullable',
-            ]);
-
-            if (Auth::user()->isFlagged) {
-                return session()->flash('error', 'Your account is flagged!');
-            }
-
             if (Auth::id() === $this->user->id) {
-                $webhook = Webhook::create([
-                    'user_id' => Auth::id(),
-                    'name' => $this->name,
-                    'product_id' => $this->product,
-                    'token' => md5(uniqid(Auth::id(), true)),
-                    'type' => $this->type,
+                $this->validate([
+                    'name' => 'required|min:2|max:20',
+                    'product' => 'nullable',
                 ]);
-                $this->name = '';
-                $this->product = '';
-                session()->flash('created', $webhook);
+    
+                if (Auth::user()->isFlagged) {
+                    return session()->flash('error', 'Your account is flagged!');
+                }
+    
+                if (Auth::id() === $this->user->id) {
+                    $webhook = Webhook::create([
+                        'user_id' => Auth::id(),
+                        'name' => $this->name,
+                        'product_id' => $this->product,
+                        'token' => md5(uniqid(Auth::id(), true)),
+                        'type' => $this->type,
+                    ]);
+                    $this->name = '';
+                    $this->product = '';
+                    session()->flash('created', $webhook);
+                } else {
+                    return session()->flash('error', 'Forbidden!');
+                }
             } else {
                 return session()->flash('error', 'Forbidden!');
             }
@@ -64,9 +68,13 @@ class Integrations extends Component
     public function deleteWebhook($id)
     {
         if (Auth::check()) {
-            $webhook = Webhook::find($id);
-            $webhook->delete();
-            $this->emit('webhookDeleted');
+            if (Auth::id() === $this->user->id) {
+                $webhook = Webhook::find($id);
+                $webhook->delete();
+                $this->emit('webhookDeleted');
+            } else {
+                return session()->flash('error', 'Forbidden!');
+            }
         } else {
             return session()->flash('error', 'Forbidden!');
         }
